Apply material styling
