+++ src/index.html
@@ -36,7 +36,6 @@
   </div>
 </nav>
 <div id="editor"></div>
-<!-- <button id="saveData">Push Data</button> -->
 </body>
 <script src="assets/bundle_ecore.js"></script>
 </html>
