+++ src/index.ts
@@ -1,13 +1,12 @@
 /* tslint:disable:no-invalid-this */
 import * as _ from 'lodash';
+import * as Ajv from 'ajv';
 import './ereference.renderer';
 import './eattribute.renderer';
 import './materialized.tree.renderer';
 import './jsoneditor';
-import { JsonEditor } from './jsoneditor';
-import { JsonForms, Style } from 'jsonforms';
-import { imageProvider, labelProvider, modelMapping } from './ecore-config';
-import * as Ajv from 'ajv';
+import { applyMaterialStyle } from './material.styling';
+import { EcoreEditor } from './ecore-editor';
 declare let $;
 export * from './jsoneditor';
 export * from './ecore-editor';
@@ -59,7 +58,7 @@ const fileInputHandler = editor => evt => {
 };
 
 window.onload = () => {
-  const editor = document.createElement('json-editor') as JsonEditor;
+  const editor = document.createElement('ecore-editor') as EcoreEditor;
 
   // create hidden file input element
   const fileInput = document.createElement('input') as HTMLInputElement;
@@ -77,77 +76,9 @@ window.onload = () => {
   uploadButton.onclick = () => {
     fileInput.click();
   };
-  xhttp2.open('GET', 'http://localhost:3001/schema.json', true);
-  // xhttp.open("GET", "http://localhost:3001/task.schema.json", true);
-  xhttp2.send();
 
   editor.data = {};
-  document.body.appendChild(editor);
+  document.getElementById('editor').appendChild(editor);
 
-  material();
-};
-
-
-// TODO
-const material = () => {
-  JsonForms.stylingRegistry.registerMany([
-    {
-      name: 'button',
-      classNames: ['btn', 'waves-effect', 'waves-light']
-    },
-    {
-      name: 'array.button',
-      classNames: ['btn-floating', 'waves-effect', 'waves-light', 'array-button']
-    },
-    {
-      name: 'array.layout',
-      classNames: ['z-depth-3']
-    },
-    {
-      name: 'group.layout',
-      classNames: ['z-depth-3']
-    },
-    {
-      name: 'group.label',
-      classNames: ['group.label']
-    },
-    {
-      name: 'collection',
-      classNames: ['collection']
-    },
-    {
-      name: 'item',
-      classNames: ['collection-item']
-    },
-    {
-      name: 'item-active',
-      classNames: ['active']
-    },
-    {
-      name: 'horizontal-layout',
-      classNames: ['row']
-    },
-    {
-      name: 'json-forms',
-      classNames: ['container']
-    }
-  ]);
-
-  JsonForms.stylingRegistry.register({
-    name: 'horizontal-layout-item',
-    classNames: childrenSize => {
-      const colSize = Math.floor(12 / childrenSize[0]);
-
-      return ['col', `s${colSize}`];
-    }
-  } as Style);
-
-  JsonForms.stylingRegistry.register({
-    name: 'vertical-layout-item',
-    classNames: ['vertical-layout-item']
-  });
-  JsonForms.stylingRegistry.deregister('select');
-
-  // init selection combo box
-  $('select').material_select();
+  applyMaterialStyle();
 };
